python/Plotly-AI/requirements.txt
dash
plotly
databricks-connect>=13.0.0
pyspark-ai